+++ api/org.integratedmodelling.klab.api/src/org/integratedmodelling/kim/api/IKimConcept.java
@@ -497,4 +497,11 @@ public interface IKimConcept extends IKimStatement {
 	 */
 	IKimConcept getAdjacent();
 
+	/**
+	 * Return a string suitable for naming a k.IM object after this concept.
+	 * 
+	 * @return
+	 */
+	String getCodeName();
+
 }